pyproject.toml
[project]
name = "discogs-finder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.2.1",
    "python3-discogs-client>=2.8",
]
